+++ zcash_client_sqlite/src/wallet/transparent.rs
@@ -3,6 +3,8 @@ use std::collections::{HashMap, HashSet};
 
 use rusqlite::OptionalExtension;
 use rusqlite::{named_params, Connection, Row};
+use zcash_client_backend::data_api::TransactionDataRequest;
+use zcash_primitives::transaction::builder::DEFAULT_TX_EXPIRY_DELTA;
 use zip32::{DiversifierIndex, Scope};
 
 use zcash_address::unified::{Encoding, Ivk, Uivk};
@@ -20,9 +22,8 @@ use zcash_primitives::{
 };
 use zcash_protocol::consensus::{self, BlockHeight};
 
-use crate::{error::SqliteClientError, AccountId, UtxoId};
-
 use super::{chain_tip_height, get_account_ids};
+use crate::{error::SqliteClientError, AccountId, TxRef, UtxoId};
 
 pub(crate) mod ephemeral;
 
@@ -227,8 +228,7 @@ pub(crate) fn get_wallet_transparent_output(
                     OR tx.expiry_height >= :mempool_height -- the spending tx has not yet expired
                  )
              )
-         )
-         ",
+         )",
     )?;
 
     let result: Result<Option<WalletTransparentOutput>, SqliteClientError> = stmt_select_utxo
@@ -429,7 +429,7 @@ pub(crate) fn add_transparent_account_balances(
 /// Marks the given UTXO as having been spent.
 pub(crate) fn mark_transparent_utxo_spent(
     conn: &rusqlite::Connection,
-    tx_ref: i64,
+    tx_ref: TxRef,
     outpoint: &OutPoint,
 ) -> Result<(), SqliteClientError> {
     let mut stmt_mark_transparent_utxo_spent = conn.prepare_cached(
@@ -441,14 +441,26 @@ pub(crate) fn mark_transparent_utxo_spent(
          AND txo.output_index = :prevout_idx
          ON CONFLICT (transparent_received_output_id, transaction_id) DO NOTHING",
     )?;
+    stmt_mark_transparent_utxo_spent.execute(named_params![
+        ":spent_in_tx": tx_ref.0,
+        ":prevout_txid": outpoint.hash().as_ref(),
+        ":prevout_idx": outpoint.n(),
+    ])?;
 
-    let sql_args = named_params![
-        ":spent_in_tx": &tx_ref,
-        ":prevout_txid": &outpoint.hash().to_vec(),
-        ":prevout_idx": &outpoint.n(),
-    ];
+    // Since we know that the output is spent, we no longer need to search for
+    // it to find out if it has been spent.
+    let mut stmt_remove_spend_detection = conn.prepare_cached(
+        "DELETE FROM transparent_spend_search_queue
+         WHERE output_index = :prevout_idx
+         AND transaction_id IN (
+            SELECT id_tx FROM transactions WHERE txid = :prevout_txid
+         )",
+    )?;
+    stmt_remove_spend_detection.execute(named_params![
+        ":prevout_txid": outpoint.hash().as_ref(),
+        ":prevout_idx": outpoint.n(),
+    ])?;
 
-    stmt_mark_transparent_utxo_spent.execute(sql_args)?;
     Ok(())
 }
 
@@ -468,11 +480,46 @@ pub(crate) fn put_received_transparent_utxo<P: consensus::Parameters>(
             output.mined_height(),
             address,
             receiving_account,
+            true,
         )
     } else {
         // The UTXO was not for any of our transparent addresses.
         Err(SqliteClientError::AddressNotRecognized(*address))
     }
+}
+
+/// Returns the vector of [`TransactionDataRequest`]s that represents the information needed by the
+/// wallet backend in order to be able to present a complete view of wallet history and memo data.
+pub(crate) fn transaction_data_requests<P: consensus::Parameters>(
+    conn: &rusqlite::Connection,
+    params: &P,
+) -> Result<Vec<TransactionDataRequest>, SqliteClientError> {
+    // We cannot construct address-based transaction data requests for the case where we cannot
+    // determine the height at which to begin, so we require that either the target height or mined
+    // height be set.
+    let mut address_request_stmt = conn.prepare_cached(
+        "SELECT ssq.address, IFNULL(t.target_height, t.mined_height)
+         FROM transparent_spend_search_queue ssq
+         JOIN transactions t ON t.id_tx = ssq.transaction_id
+         WHERE t.target_height IS NOT NULL
+         OR t.mined_height IS NOT NULL",
+    )?;
+
+    let result = address_request_stmt
+        .query_and_then([], |row| {
+            let address = TransparentAddress::decode(params, &row.get::<_, String>(0)?)?;
+            let block_range_start = BlockHeight::from(row.get::<_, u32>(1)?);
+            Ok::<TransactionDataRequest, SqliteClientError>(
+                TransactionDataRequest::SpendsFromAddress {
+                    address,
+                    block_range_start,
+                    block_range_end: Some(block_range_start + DEFAULT_TX_EXPIRY_DELTA + 1),
+                },
+            )
+        })?
+        .collect::<Result<Vec<_>, _>>()?;
+
+    Ok(result)
 }
 
 pub(crate) fn get_transparent_address_metadata<P: consensus::Parameters>(
@@ -573,6 +620,7 @@ pub(crate) fn find_account_for_transparent_address<P: consensus::Parameters>(
 ///
 /// `output_height` may be None if this is an ephemeral output from a
 /// transaction we created, that we do not yet know to have been mined.
+#[allow(clippy::too_many_arguments)]
 pub(crate) fn put_transparent_output<P: consensus::Parameters>(
     conn: &rusqlite::Connection,
     params: &P,
@@ -581,6 +629,7 @@ pub(crate) fn put_transparent_output<P: consensus::Parameters>(
     output_height: Option<BlockHeight>,
     address: &TransparentAddress,
     receiving_account: AccountId,
+    known_unspent: bool,
 ) -> Result<UtxoId, SqliteClientError> {
     let output_height = output_height.map(u32::from);
 
@@ -613,6 +662,40 @@ pub(crate) fn put_transparent_output<P: consensus::Parameters>(
         |row| row.get::<_, i64>(0),
     )?;
 
+    let spent_height = conn
+        .query_row(
+            "SELECT t.mined_height
+             FROM transactions t
+             JOIN transparent_received_output_spends ts ON ts.transaction_id = t.id_tx
+             JOIN transparent_received_outputs tro ON tro.id = ts.transparent_received_output_id
+             WHERE tro.transaction_id = :transaction_id
+             AND tro.output_index = :output_index",
+            named_params![
+                ":transaction_id": id_tx,
+                ":output_index": &outpoint.n(),
+            ],
+            |row| {
+                row.get::<_, Option<u32>>(0)
+                    .map(|o| o.map(BlockHeight::from))
+            },
+        )
+        .optional()?
+        .flatten();
+
+    // The max observed unspent height is either the spending transaction's mined height - 1, or
+    // the current chain tip height if the UTXO was received via a path that confirmed that it was
+    // unspent, such as by querying the UTXO set of the network.
+    let max_observed_unspent = match spent_height {
+        Some(h) => Some(h - 1),
+        None => {
+            if known_unspent {
+                chain_tip_height(conn)?
+            } else {
+                None
+            }
+        }
+    };
+
     let mut stmt_upsert_transparent_output = conn.prepare_cached(
         "INSERT INTO transparent_received_outputs (
             transaction_id, output_index,
@@ -622,14 +705,14 @@ pub(crate) fn put_transparent_output<P: consensus::Parameters>(
         VALUES (
             :transaction_id, :output_index,
             :account_id, :address, :script,
-            :value_zat, :height
+            :value_zat, :max_observed_unspent_height
         )
         ON CONFLICT (transaction_id, output_index) DO UPDATE
         SET account_id = :account_id,
             address = :address,
             script = :script,
             value_zat = :value_zat,
-            max_observed_unspent_height = :height
+            max_observed_unspent_height = IFNULL(:max_observed_unspent_height, max_observed_unspent_height)
         RETURNING id",
     )?;
 
@@ -640,12 +723,45 @@ pub(crate) fn put_transparent_output<P: consensus::Parameters>(
         ":address": &address.encode(params),
         ":script": &txout.script_pubkey.0,
         ":value_zat": &i64::from(Amount::from(txout.value)),
-        ":height": output_height,
+        ":max_observed_unspent_height": max_observed_unspent.map(u32::from),
     ];
 
     let utxo_id = stmt_upsert_transparent_output
         .query_row(sql_args, |row| row.get::<_, i64>(0).map(UtxoId))?;
     Ok(utxo_id)
+}
+
+/// Adds a request to retrieve transactions involving the specified address to the transparent
+/// spend search queue. Note that such requests are _not_ for data related to `tx_ref`, but instead
+/// a request to find where the UTXO with the outpoint `(tx_ref, output_index)` is spent.
+///
+/// ### Parameters
+/// - `receiving_address`: The address that received the UTXO.
+/// - `tx_ref`: The transaction in which the UTXO was received.
+/// - `output_index`: The index of the output within `vout` of the specified transaction.
+pub(crate) fn queue_transparent_spend_detection<P: consensus::Parameters>(
+    conn: &rusqlite::Transaction<'_>,
+    params: &P,
+    receiving_address: TransparentAddress,
+    tx_ref: TxRef,
+    output_index: u32,
+) -> Result<(), SqliteClientError> {
+    let mut stmt = conn.prepare_cached(
+        "INSERT INTO transparent_spend_search_queue
+         (address, transaction_id, output_index)
+         VALUES
+         (:address, :transaction_id, :output_index)
+         ON CONFLICT (transaction_id, output_index) DO NOTHING",
+    )?;
+
+    let addr_str = receiving_address.encode(params);
+    stmt.execute(named_params! {
+        ":address": addr_str,
+        ":transaction_id": tx_ref.0,
+        ":output_index": output_index
+    })?;
+
+    Ok(())
 }
 
 #[cfg(test)]